# Calculating VAVAILABILITY