… .travis.yml
… tests/VObject/Component/AvailableTest.php
… tests/VObject/Component/VAvailabilityTest.php
… tests/VObject/FreeBusyDataTest.php
… tests/VObject/FreeBusyGeneratorTest.php
… tests/phpunit.xml