@evert 2015-07-17
Not gonna bother with making it faster. It's much better this way and we already got major speedups in the 4.x branch, I think it's within acceptable limits.

I'm also not reintegrating FreeBusyData into FreeBusyGenerator. It's a bit unfortunate to have a separate class, but it does make things a lot more clear.

So when I fixed the broken tests, this is ready for merging.